@@ -10,23 +10,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.1.5",
+    "@angular/animations": "^15.2.0",
     "@angular/common": "^15.2.0",
     "@angular/compiler": "^15.2.0",
-    "@angular/core": "^15.1.4",
-    "@angular/forms": "^15.1.5",
-    "@angular/material": "^15.1.4",
-    "@angular/platform-browser": "^15.1.5",
-    "@angular/platform-browser-dynamic": "^15.1.4",
+    "@angular/core": "^15.2.0",
+    "@angular/forms": "^15.2.0",
+    "@angular/material": "^15.2.1",
+    "@angular/platform-browser": "^15.2.0",
+    "@angular/platform-browser-dynamic": "^15.2.0",
     "@angular/router": "^15.2.0",
     "flowbite": "^1.6.3",
+    "primeicons": "^6.0.1",
+    "primeng": "^15.2.0",
     "rxjs": "~7.5.7",
     "uuid": "^9.0.0",
     "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.5",
-    "@angular/cdk": "^15.1.4",
+    "@angular-devkit/build-angular": "^15.2.0",
+    "@angular/cdk": "^15.2.1",
     "@angular/cli": "^15.2.0",
     "@angular/compiler-cli": "^15.2.0",
     "@types/jasmine": "~4.3.1",
@@ -39,7 +41,7 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
     "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.6",
+    "tailwindcss": "^3.2.7",
     "tslib": "^2.5.0",
     "typescript": "~4.9.5"
   }
